404 에러 코드 처리

+++ src/main/java/io/inforzia/todolist/TodoController.java
@@ -9,6 +9,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @Slf4j
@@ -21,7 +22,9 @@ public class TodoController {
 
     @ApiOperation(value = "TODO 생성", notes = "해야할 일과 완료 여부를 작성하여 저장합니다.")
     @PostMapping(path = "/todo", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<TodoItem> postTodo(@RequestHeader(value="authorization") String author, @RequestBody TodoItem todoItem) {
+    public ResponseEntity<TodoItem> postTodo(@RequestHeader(value="authorization") String author, @RequestBody TodoItem todoItem) throws RestException{
+        if(todoItem.getItem()==null||todoItem.getCompleted()==null)
+            throw new RestException();
         todoItem.setUser(author);
         TodoItem savedTodo = todoItemRepository.save(todoItem);
         return ResponseEntity.ok(savedTodo);
@@ -29,7 +32,9 @@ public class TodoController {
 
     @ApiOperation(value = "TODO 수정", notes = "해야할 일과 완료 여부를 수정하여 저장합니다.")
     @PutMapping(path = "/todo/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public String putTodo(@PathVariable Integer id, @RequestBody TodoItem newTodoItem) {
+    public String putTodo(@PathVariable Integer id, @RequestBody TodoItem newTodoItem) throws RestException{
+        if(!todoItemRepository.existsById(id))
+            throw new RestException();
         todoItemRepository.findById(id)
                 .map(todoItem -> {
                     todoItem.setItem(newTodoItem.getItem());
@@ -39,14 +44,16 @@ public class TodoController {
                 .orElseGet(() -> {
                     newTodoItem.
                             setId(id);
-                    return todoItemRepository.save(newTodoItem); //TODO 예외던지기
+                    return todoItemRepository.save(newTodoItem);
                 });
         return id + "번 TODO 수정";
     }
 
     @ApiOperation(value = "TODO의 완료여부 변경", notes = "해당 TODO의 완료 여부를 completed로 변경합니다.")
     @PatchMapping(path = "/todo/{id}/complete")
-    public String patchTodoCompleted(@PathVariable(value = "id") Integer id) {
+    public String patchTodoCompleted(@PathVariable(value = "id") Integer id) throws RestException{
+        if(!todoItemRepository.existsById(id))
+            throw new RestException();
         todoItemRepository.findById(id)
                 .map(todoItem -> {
                     todoItem.setCompleted(true);
@@ -57,7 +64,9 @@ public class TodoController {
 
     @ApiOperation(value = "TODO의 완료여부 변경", notes = "해당 TODO의 완료 여부를 incompleted로 변경합니다.")
     @PatchMapping(path = "/todo/{id}/incomplete")
-    public String patchTodoIncompleted(@PathVariable(value = "id") Integer id) {
+    public String patchTodoIncompleted(@PathVariable(value = "id") Integer id) throws RestException {
+        if(!todoItemRepository.existsById(id))
+            throw new RestException();
         todoItemRepository.findById(id)
                 .map(todoItem -> {
                     todoItem.setCompleted(false);
@@ -68,7 +77,9 @@ public class TodoController {
 
     @ApiOperation(value = "TODO 삭제", notes = "해당 TODO를 삭제합니다.")
     @DeleteMapping(path = "/todo/{id}")
-    public String deleteTodo(@PathVariable Integer id) {
+    public String deleteTodo(@PathVariable Integer id) throws RestException {
+        if(!todoItemRepository.existsById(id))
+            throw new RestException();
         todoItemRepository.deleteById(id);
         return id + "번 TODO 삭제";
     }
@@ -83,9 +94,10 @@ public class TodoController {
 
     @ApiOperation(value = "TODO 조회", notes = "해당 TODO의 정보를 조회합니다.")
     @GetMapping(path = "/todo/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public TodoItem getTodo(@PathVariable Integer id) {
-        return todoItemRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("illegal argument :" + id));
+    public Optional<TodoItem> getTodo(@PathVariable Integer id) throws RestException{
+        if(!todoItemRepository.existsById(id))
+            throw new RestException();
+        return todoItemRepository.findById(id);
     }
 
     @ApiOperation(value = "TODO 조회", notes = "모든 TODO를 조회하거나 완료 여부로 조회합니다.")
